Refactoring - Rename Method: The method calculateProbability is moved from the class SpatialProbabilityImpl.java to the class SpatialFilter.java as the method has been used more in SpatialFilter.java than in SpatialProbabilityImpl.java
